+++ influxdb2_client/src/api/mod.rs
@@ -1,4 +1,5 @@
 //! InfluxDB v2.0 Client API
+pub mod buckets;
 pub mod query;
 pub mod ready;
 pub mod setup;
